refactor: compose on channel page (#1190)
* refactor: open compose modal with channel

* refactor: code style

* refactor: useUpdateCurrentVisitingChannel

* refactor: remove double-checking
