+++ src/components/ComposeButton/ComposeButtonForMobile.tsx
@@ -5,6 +5,7 @@ import ReplyIcon from '@/assets/reply.svg';
 import { ClickableButton } from '@/components/ClickableButton.js';
 import { isRoutePathname } from '@/helpers/isRoutePathname.js';
 import { narrowToSocialSource } from '@/helpers/narrowSource.js';
+import { useCurrentVisitingChannel } from '@/hooks/useCurrentVisitingChannel.js';
 import { useCurrentVisitingPost } from '@/hooks/useCurrentVisitingPost.js';
 import { useIsLogin } from '@/hooks/useIsLogin.js';
 import { ComposeModalRef } from '@/modals/controls.js';
@@ -20,6 +21,7 @@ export function ComposeButtonForMobile() {
     const isLogin = useIsLogin();
     const isCurrentLogin = useIsLogin(currentSocialSource);
     const currentPost = useCurrentVisitingPost();
+    const currentChannel = useCurrentVisitingChannel();
 
     if (!isLogin) return null;
     if (isPostPage && !isCurrentLogin) return null;
@@ -30,7 +32,8 @@ export function ComposeButtonForMobile() {
             onClick={() => {
                 ComposeModalRef.open({
                     type: isPostPage ? 'reply' : 'compose',
-                    post: isPostPage ? currentPost : undefined,
+                    post: currentPost,
+                    channel: currentChannel,
                 });
             }}
         >
